v3.5.0

Because of the large structural transition that happened from 3.4 to 3.5
the decision was made to use a merge.

git merge --strategy-option theirs v3.5.0

+++ tests/e2e/etcd_process.go
@@ -19,14 +19,16 @@ import (
 	"net/url"
 	"os"
 
-	"go.etcd.io/etcd/pkg/expect"
-	"go.etcd.io/etcd/pkg/fileutil"
+	"go.etcd.io/etcd/client/pkg/v3/fileutil"
+	"go.etcd.io/etcd/pkg/v3/expect"
+	"go.uber.org/zap"
 )
 
 var (
-	etcdServerReadyLines = []string{"enabled capabilities for version", "published"}
+	etcdServerReadyLines = []string{"ready to serve client requests"}
 	binPath              string
 	ctlBinPath           string
+	utlBinPath           string
 )
 
 // etcdProcess is a process that serves etcd requests.
@@ -50,6 +52,7 @@ type etcdServerProcess struct {
 }
 
 type etcdServerProcessConfig struct {
+	lg       *zap.Logger
 	execPath string
 	args     []string
 	tlsArgs  []string
@@ -70,7 +73,7 @@ type etcdServerProcessConfig struct {
 
 func newEtcdServerProcess(cfg *etcdServerProcessConfig) (*etcdServerProcess, error) {
 	if !fileutil.Exist(cfg.execPath) {
-		return nil, fmt.Errorf("could not find etcd binary")
+		return nil, fmt.Errorf("could not find etcd binary: %s", cfg.execPath)
 	}
 	if !cfg.keepDataDir {
 		if err := os.RemoveAll(cfg.dataDirPath); err != nil {
@@ -88,23 +91,34 @@ func (ep *etcdServerProcess) Start() error {
 	if ep.proc != nil {
 		panic("already started")
 	}
-	proc, err := spawnCmd(append([]string{ep.cfg.execPath}, ep.cfg.args...))
+	ep.cfg.lg.Info("starting server...", zap.String("name", ep.cfg.name))
+	proc, err := spawnCmdWithLogger(ep.cfg.lg, append([]string{ep.cfg.execPath}, ep.cfg.args...))
 	if err != nil {
 		return err
 	}
 	ep.proc = proc
-	return ep.waitReady()
+	err = ep.waitReady()
+	if err == nil {
+		ep.cfg.lg.Info("started server.", zap.String("name", ep.cfg.name))
+	}
+	return err
 }
 
 func (ep *etcdServerProcess) Restart() error {
+	ep.cfg.lg.Info("restaring server...", zap.String("name", ep.cfg.name))
 	if err := ep.Stop(); err != nil {
 		return err
 	}
 	ep.donec = make(chan struct{})
-	return ep.Start()
+	err := ep.Start()
+	if err == nil {
+		ep.cfg.lg.Info("restared server", zap.String("name", ep.cfg.name))
+	}
+	return err
 }
 
 func (ep *etcdServerProcess) Stop() (err error) {
+	ep.cfg.lg.Info("stoping server...", zap.String("name", ep.cfg.name))
 	if ep == nil || ep.proc == nil {
 		return nil
 	}
@@ -117,18 +131,24 @@ func (ep *etcdServerProcess) Stop() (err error) {
 	ep.donec = make(chan struct{})
 	if ep.cfg.purl.Scheme == "unix" || ep.cfg.purl.Scheme == "unixs" {
 		err = os.Remove(ep.cfg.purl.Host + ep.cfg.purl.Path)
-		if err != nil {
+		if err != nil && !os.IsNotExist(err) {
 			return err
 		}
 	}
+	ep.cfg.lg.Info("stopped server.", zap.String("name", ep.cfg.name))
 	return nil
 }
 
 func (ep *etcdServerProcess) Close() error {
+	ep.cfg.lg.Info("closing server...", zap.String("name", ep.cfg.name))
 	if err := ep.Stop(); err != nil {
 		return err
 	}
-	return os.RemoveAll(ep.cfg.dataDirPath)
+	if !ep.cfg.keepDataDir {
+		ep.cfg.lg.Info("removing directory", zap.String("data-dir", ep.cfg.dataDirPath))
+		return os.RemoveAll(ep.cfg.dataDirPath)
+	}
+	return nil
 }
 
 func (ep *etcdServerProcess) WithStopSignal(sig os.Signal) os.Signal {
